+++ product-service/handlers/createProduct.js
@@ -1,5 +1,5 @@
 import { createDBProduct } from '../services/productService';
-import { requestHandler } from '../../utils/http.util';
+import { requestHandler } from '../utils/http.util';
 
 export const createProduct = (event) => requestHandler(event, async (event) => {
   const product = event?.body && JSON.parse(event.body);
