+++ product-service/handlers/getProductById.js
@@ -1,5 +1,5 @@
 import { getDBProductById } from '../services/productService';
-import { requestHandler } from '../../utils/http.util';
+import { requestHandler } from '../utils/http.util';
 
 export const getProductById = (event) => requestHandler(event, async (e) => {
   const { productId } = e.pathParameters;
