+++ product-service/handlers/getProducts.js
@@ -1,5 +1,5 @@
 import { getDBProducts } from '../services/productService';
-import { requestHandler } from '../../utils/http.util';
+import { requestHandler } from '../utils/http.util';
 
 export const getProducts = (event) => requestHandler(event, async (e) => {
     return await getDBProducts();
